+++ grafton-visca-macros/src/value_macros.rs
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: MIT OR Apache-2.0
 //
 // Copyright (c) 2024 Grafton Machine Shed <team@grafton.ai>